Moved the src::compiler::llvm submodule to src::llvm, now that there is only LLVM for the compiler backend there is no need for the compiler module for organization;
